web-animations/animation-timeline/animation-timeline-play-timeline-current-time-test.html

<!DOCTYPE html>
<html>
<head>
<title>Animation timeline play() method test. Test current time of newly-created AnimationPlayer</title>
<meta name="assert" content="The currentTime attribute of the AnimationPlayer object is a calculated value described in section '3.5.1 The current time of a player'">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#the-animationtimeline-interface">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    var player = newAnimationPlayer(doc);
	var currentTime = calculateCurrentTime(player.timeline.currentTime, getAdjustedTime(player),
	    player.playbackRate, getHoldTime(player));
    
    assert_equals(player.currentTime, currentTime,
        'If hold time of AnimationPlayer object is null, the current time ' +
        'should be current time = (timeline time - adjusted start time) * playback rate');
}, 'Current time of Animation Player calculated is ' +
    'current time = (timeline time - adjusted start time) × playback rate. Player is not paused');


test(function() {
    var doc = newHTMLDocument();
    var player = newAnimationPlayer(doc);
    player.pause();
	var currentTime = calculateCurrentTime(player.timeline.currentTime, getAdjustedTime(player),
        player.playbackRate, getHoldTime(player));
    
    assert_equals(player.currentTime, currentTime,
        'If hold time of AnimationPlayer object is not null, the current time ' +
        'should be the hold time');
}, 'Current time of Animation Player calculated is a hold time when player is paused');

</script>
</body>
</html>
